+++ eggs/VanillaIceCream/src/main/java/com/android_v/egg/landroid/Universe.kt
@@ -295,6 +295,10 @@ open class Universe(val namer: Namer, randomSeed: Long) : Simulator(randomSeed) 
                     //                        vDiff < 100f
                     ) {
                         val landing = Landing(ship, planet, a, namer.describeActivity(rng, planet))
+                        if (ship.thrust != Vec2.Zero) {
+                            // kill the power
+                            ship.thrust = Vec2.Zero
+                        }
                         ship.landing = landing
                         ship.velocity = planet.velocity
                         add(landing)